Change breeding to be more realistic
Before, if one male organism was present, all female organisms would be
impregnated. This is now changed so that each male has a limit on the
amount of females it will impregnate in a single turn.
A limit on the amount of times the male will try is also enforced.